requirements.txt
Flask==2.3.2
fido2==1.2.0
flask-cors
flask-sqlalchemy
psycopg2-binary
gunicorn
python-dotenv